requirements.txt
# Data processing and numerical computations
numpy==1.22.4
opencv-python-headless==4.5.5.64
Pillow==9.1.1

# Deep learning frameworks
keras==2.9.0
tensorflow==2.9.1
torch==1.11.0
torchvision==0.12.0

# Transformer models
transformers==4.19.2

# Utilities
pyyaml==6.0
imutils==0.5.4
